Extract mdast and hast helpers into src/markdown.ts
